+++ Source/Dispatcher.swift
@@ -26,20 +26,35 @@ open class Dispatcher {
     /// Dispatcher queue priority
     private final var priority = Priority.operation
 
+    /// Main queue dispatcher
+    open static var main: Dispatcher { return MainDispatcher() }
+    /// Current queue dispatcher
+    open static var current: Dispatcher { return CurrentDispatcher() }
+    /// Dispatcher with default quality of service
+    open static var `default`: Dispatcher { return Dispatcher() }
+    /// Dispatcher with user interactive quality of service
+    open static var userInteractive: Dispatcher { return Dispatcher(name: nil, qos: .userInteractive) }
+    /// Dispatcher with user initiated quality of service
+    open static var userInitiated: Dispatcher { return Dispatcher(name: nil, qos: .userInitiated) }
+    /// Dispatcher with utility quality of service
+    open static var utility: Dispatcher { return Dispatcher(name: nil, qos: .utility) }
+    /// Dispatcher with background quality of service
+    open static var background: Dispatcher { return Dispatcher(name: nil, qos: .background) }
+
     /// Dispatcher instance with default OperationQueue
     public convenience init() {
-        self.init(name: nil, qos: nil, maxConcurrentOperationCount: OperationQueue.defaultMaxConcurrentOperationCount)
+        self.init(name: nil, qos: .default)
     }
 
     /// Dispatcher instance with custom OperationQueue
-    public init(name: String?, qos: QualityOfService?, maxConcurrentOperationCount: Int) {
+    public init(name: String?, qos: QualityOfService?, maxConcurrentOperationCount: Int? = nil) {
         operationQueue.name = name ?? UUID().uuidString
         operationQueue.qualityOfService = qos ?? .default
-        operationQueue.maxConcurrentOperationCount = maxConcurrentOperationCount
+        operationQueue.maxConcurrentOperationCount = maxConcurrentOperationCount ?? OperationQueue.defaultMaxConcurrentOperationCount
     }
 
     /// Dispatcher instance with custom DispatchQueue
-    public init(label: String?, qos: DispatchQoS?, attributes: DispatchQueue.Attributes?, autoreleaseFrequency: DispatchQueue.AutoreleaseFrequency?, target: DispatchQueue?) {
+    public init(label: String?, qos: DispatchQoS?, attributes: DispatchQueue.Attributes? = nil, autoreleaseFrequency: DispatchQueue.AutoreleaseFrequency? = nil, target: DispatchQueue? = nil) {
         dispatchQueue = DispatchQueue(label: label ?? UUID().uuidString, qos: qos ?? .default, attributes: attributes ?? .concurrent, autoreleaseFrequency: autoreleaseFrequency ?? .inherit, target: target)
         priority = .dispatch
     }
@@ -79,6 +94,21 @@ open class Dispatcher {
         }
         execute(operations, waitUntilFinished: waitUntilFinished)
         return operations
+    }
+
+    /// Execute block in DispatchQueue after delay
+    open func execute(after delay: TimeInterval, block: @escaping () -> Void) {
+        dispatchQueue.asyncAfter(deadline: .now() + delay, execute: block)
+    }
+
+    /// Execute block in DispatchQueue using custom DispatchWorkItem instance after delay
+    open func execute(after delay: TimeInterval, qos: DispatchQoS?, flags: DispatchWorkItemFlags?, block: @escaping @convention(block) () -> ()) {
+        dispatchQueue.asyncAfter(deadline: .now() + delay, qos: qos ?? .default, flags: flags ?? .assignCurrentContext, execute: block)
+    }
+
+    /// Execute DispatchWorkItem instance in DispatchQueue after delay
+    open func execute(after delay: TimeInterval, work: DispatchWorkItem) {
+        dispatchQueue.asyncAfter(deadline: .now() + delay, execute: work)
     }
 
     /// Execute block in DispatchQueue using custom DispatchWorkItem instance