+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Release 0.6.0
+
+- [x] Execute and cancel after delay
+- [x] Kommand cancellation fix
+
 # Release 0.5.0
 
 - [x] Concurrent/Sequential Kommands execution refactor